Fix bugs in setup.sh

+++ setup.sh
@@ -6,7 +6,6 @@ git submodule update
 shopt -s dotglob
 cd ~
 
-
 for f in $OLDPWD/*; do
     O="$(basename $f)"
     if [[ "$O" == "." ]]; then continue; fi
@@ -14,6 +13,7 @@ for f in $OLDPWD/*; do
     if [[ "$O" == ".git" ]]; then continue; fi
     if [[ "$O" == ".gitignore" ]]; then continue; fi
     if [[ "$O" == ".gitmodules" ]]; then continue; fi
+    if [[ "$O" == "setup.sh" ]]; then continue; fi
     if [[ "$O" == "README.md" ]]; then continue; fi
 
     rm "$O" >/dev/null 2>&1
@@ -24,4 +24,4 @@ shopt -u dotglob
 cd - >/dev/null 2>&1
 
 #Setup vim
-git clone --recursive https://github.com/rgooler/.vim ${HOMEDIR}/.vim
+git clone --recursive https://github.com/rgooler/.vim
